# api-contract-v1.yaml
openapi: 3.1.0
info:
  title: Event Face Matching API
  version: 1.0.0
  description: |
    Phase 1 contract — single studio, no billing. Multi-tenant structures are
    present but only one studio row exists.

    ## Who owns what
    Backend (routes, DB, worker) is implemented by the studio owner.
    Frontend (three React apps) is implemented against this spec via a mock server.
    Neither side blocks the other. When the real server is live, only the base URL changes.

    ## Two separate auth contexts — do not mix them
    - **Staff** (studio owners, photographers, platform admins) use
      `Authorization: Bearer <access_token>`.
    - **Guests** have no account. They use `X-Guest-Session: <session_token>`,
      scoped to exactly one event, and can reach only the `/g/*` routes.

    Keeping these in different headers is deliberate: it makes it structurally
    hard to accidentally accept a guest token on a staff route.

    ## photos.status state machine
    ```
    pending ──► uploaded ──► processing ──► done
                                 │
                                 └────────► failed ──► (retry) ──► processing
    ```
    - `pending`    row created, presigned URL issued, bytes not yet confirmed
    - `uploaded`   client confirmed the PUT succeeded; job enqueued
    - `processing` a worker has claimed the job
    - `done`       faces extracted, embedded, stored, matched against guests
    - `failed`     gave up after 5 attempts; `error_code` explains why

...

...
servers:
...
security:
  - staffAuth: []
paths:
...
  /studio/members:
    get:
      tags: [members]
      summary: Everyone who belongs to the caller's studio
      description: |
        Owners and photographers alike, including invitations that have not been
        accepted yet. Scoped to the caller's studio by row-level security, so
        there is no studio id in the path.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/StudioMember' }

    post:
      tags: [members]
      summary: Create an account for a photographer
      description: |
        No email is sent and no invitation is accepted. The server generates an
        initial password and returns it **once**; the owner passes it on over
        WhatsApp. Same pattern the platform uses to create a studio.

        `initial_password` appears in this response and nowhere else, ever. The
        account is created with `must_change_password` set, so the photographer
        replaces it the first time they sign in.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, role]
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 40
                  pattern: "^[a-z0-9._-]+$"
                name: { type: string, maxLength: 120 }
                email:
                  type: string
                  format: email
                  description: Optional. Enables self-service reset later.
                role: { type: string, enum: [photographer, owner] }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StudioMember'
                  - type: object
                    required: [initial_password]
                    properties:
                      initial_password:
                        type: string
                        description: Shown once. Never retrievable again.
        '409':
          description: That username is taken
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: Caller is not an owner
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /studio/members/{user_id}:
    parameters:
      - name: user_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [members]
      summary: Remove someone from the studio
      description: |
        Revokes access. Photographs they already uploaded stay with the event,
        because they belong to the studio and not to the person who sent them.

        The last remaining owner cannot be removed, otherwise the studio becomes
        unreachable.
      responses:
        '204':
          description: Removed
        '409':
          description: Would leave the studio with no owner
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                error:
                  code: LAST_OWNER
                  message: A studio must always have at least one owner.

  # ───────────────────────────── lifecycle ─────────────────────────────
...
  /studio/members/{user_id}/reset-password:
    parameters:
      - name: user_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [members]
      summary: Issue a new initial password for a photographer
      description: |
        Owner only. Returns the new password once and revokes every refresh
        token that person holds. Revocation is the point: resetting a password
        because somebody else has it, while leaving their sessions alive,
        changes nothing.
      responses:
        '200':
          description: Reset
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StudioMember'
                  - type: object
                    required: [initial_password]
                    properties:
                      initial_password: { type: string }

  # ───────────────────────────── webhooks ─────────────────────────────
...
components:
  securitySchemes:
    staffAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Studio owners, photographers and platform admins.
    guestAuth:
      type: apiKey
      in: header
      name: X-Guest-Session
      description: Opaque session token, scoped to one event, valid for `/g/*` only.
...
  responses:
...
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable identifier. Frontend switches on this, never on `message`.
            message:
              type: string
              description: Human-readable text, safe to show a user.
            details:
              type: object
              additionalProperties: true
...
    StudioMember:
      type: object
      required: [user_id, username, role, status, photos_uploaded]
      properties:
        user_id: { type: string, format: uuid }
        username: { type: string }
        email: { type: [string, 'null'] }
        name: { type: [string, 'null'] }
        role: { type: string, enum: [owner, photographer] }
        status:
          type: string
          enum: [active, pending_first_login]
          description: "`pending_first_login` until they sign in and set their own password."

        photos_uploaded:
          type: integer
          description: Across all of this studio's events. Shows who is actually working.
        last_active_at: { type: [string, 'null'], format: date-time }
        invited_at: { type: string, format: date-time }
...